Changed deprecated cuda lite runtime test_packet with all new test_tile_info that uses bsg_printf
