--- cl_manycore/regression/cuda/test_packet.c
@@ -1,105 +1,0 @@
-#include "test_packet.h"
-
-#define TEST_NAME "test_packet"
-#define ALLOC_NAME "default_allocator"
-
-/*!
- * Runs a packet kernel on a 2x2 tile group. 
- * This tests uses the software/spmd/bsg_cuda_lite_runtime/packet/ Manycore binary in the BSG Manycore github repository.  
-*/
-
-int kernel_packet () {
-	bsg_pr_test_info("Running the CUDA Packet Kernel on a 2x2 tile group.\n\n");
-	int rc;
-
-
-	/*****************************************************************************************************************
-	* Define path to binary.
-	* Initialize device, load binary and unfreeze tiles.
-	******************************************************************************************************************/
-	hb_mc_device_t device;
-	rc = hb_mc_device_init(&device, TEST_NAME, 0);
-	if (rc != HB_MC_SUCCESS) { 
-		bsg_pr_err("failed to initialize device.\n");
-		return rc;
-	}
-
-
-	char* elf = BSG_STRINGIFY(BSG_MANYCORE_DIR) "/software/spmd/bsg_cuda_lite_runtime" "/packet/main.riscv";
-	rc = hb_mc_device_program_init(&device, elf, ALLOC_NAME, 0);
-	if (rc != HB_MC_SUCCESS) { 
-		bsg_pr_err("failed to initialize program.\n");
-		return rc;
-	}
-
-
-	/*****************************************************************************************************************
-	* Define grid_dim_x/y: total number of tile groups
-	* Define tg_dim_x/y: number of tiles in each tile group
-	* Calculate grid_dim_x/y: number of tile groups needed based on block_size_x/y
-	******************************************************************************************************************/
-	hb_mc_dimension_t grid_dim = { .x = 1, .y = 1};
-	hb_mc_dimension_t tg_dim = { .x = 2, .y = 2}; 
-
-
-	/*****************************************************************************************************************
-	* Prepare list of input arguments for kernel.
-	******************************************************************************************************************/
-	int argv[1];
-
-
-	/*****************************************************************************************************************
-	* Enquque grid of tile groups, pass in grid and tile group dimensions, kernel name, number and list of input arguments
-	******************************************************************************************************************/
-	rc =hb_mc_application_init (&device, grid_dim, tg_dim, "kernel_packet", 0, argv);
-	if (rc != HB_MC_SUCCESS) { 
-		bsg_pr_err("failed to initialize grid.\n");
-		return rc;
-	}
-	
-
-	/*****************************************************************************************************************
-	* Launch and execute all tile groups on device and wait for all to finish. 
-	******************************************************************************************************************/
-	rc = hb_mc_device_tile_groups_execute(&device);
-	if (rc != HB_MC_SUCCESS) { 
-		bsg_pr_err("failed to execute tile groups.\n");
-		return rc;
-	}
-	
-
-	/*****************************************************************************************************************
-	* Freeze the tiles and memory manager cleanup. 
-	******************************************************************************************************************/
-	rc = hb_mc_device_finish(&device); /* freeze the tiles and memory manager cleanup */
-	if (rc != HB_MC_SUCCESS) { 
-		bsg_pr_err("failed to de-initialize device.\n");
-		return rc;
-	}
-
-
-	return HB_MC_SUCCESS;
-}
-
-#ifdef COSIM
-void test_main(uint32_t *exit_code) {	
-#ifdef VCS
-	svScope scope;
-	scope = svGetScopeFromName("tb");
-	svSetScope(scope);
-#endif
-	bsg_pr_test_info("test_packet Regression Test (COSIMULATION)\n");
-	int rc = kernel_packet();
-	*exit_code = rc;
-	bsg_pr_test_pass_fail(rc == HB_MC_SUCCESS);
-	return;
-}
-#else
-int main() {
-	bsg_pr_test_info("test_packet Regression Test (F1)\n");
-	int rc = kernel_packet();
-	bsg_pr_test_pass_fail(rc == HB_MC_SUCCESS);
-	return rc;
-}
-#endif
-
