+++ cl_manycore/regression/cuda/test_tile_info.c
@@ -1,0 +1,117 @@
+/******************************************************************************/
+/* Runs the tile info  printing kernel on a 2x2 tile group.                   */
+/* Each tile prints all operation symbols including __bsg_x/y, __bsg_id,      */
+/* __bsg_grp_org_x/y, __bsg_tg_id_x/y, __bsg_grid_dim_x/y.                    */
+/* Grid dimensions are prefixed at 1x1.                                       */
+/* This tests uses the software/spmd/bsg_cuda_lite_runtime/tile_info/         */
+/* manycore binary in the BSG Manycore repository.                            */
+/******************************************************************************/
+
+
+#include "test_tile_info.h"
+
+#define TEST_NAME "test_tile_info"
+#define ALLOC_NAME "default_allocator"
+
+
+int kernel_tile_info () {
+	bsg_pr_test_info("Running the CUDA Tile Info Printing Kernel "
+                         "on a grid of 2x2 tile groups.\n\n");
+	int rc;
+
+	srand(time); 
+
+
+
+        /**********************************************************************/
+        /* Define path to binary.                                             */
+        /* Initialize device, load binary and unfreeze tiles.                 */
+        /**********************************************************************/
+	hb_mc_device_t device;
+	rc = hb_mc_device_init(&device, TEST_NAME, 0);
+	if (rc != HB_MC_SUCCESS) { 
+		bsg_pr_err("failed to initialize device.\n");
+		return rc;
+	}
+
+	char* elf = BSG_STRINGIFY(BSG_MANYCORE_DIR) "/software/spmd/bsg_cuda_lite_runtime"
+                                                    "/tile_info/main.riscv";
+	rc = hb_mc_device_program_init(&device, elf, ALLOC_NAME, 0);
+	if (rc != HB_MC_SUCCESS) { 
+		bsg_pr_err("failed to initialize program.\n");
+		return rc;
+	}
+
+
+        /**********************************************************************/
+	/* Define block_size_x/y: amount of work for each tile group          */
+	/* Define tg_dim_x/y: number of tiles in each tile group              */
+	/* Calculate grid_dim_x/y: number of                                  */
+        /* tile groups needed based on block_size_x/y                         */
+        /**********************************************************************/
+	hb_mc_dimension_t tg_dim = { .x = 2, .y = 2 }; 
+
+	hb_mc_dimension_t grid_dim = { .x = 1, .y = 1 };
+
+
+        /**********************************************************************/
+	/* Prepare list of input arguments for kernel.                        */
+        /**********************************************************************/
+	int argv[1] = {};
+
+
+        /**********************************************************************/
+	/* Enquque grid of tile groups, pass in grid and tile group dimensions*/
+        /* kernel name, number and list of input arguments                    */
+        /**********************************************************************/
+	rc = hb_mc_application_init (&device, grid_dim, tg_dim, "kernel_tile_info", 0, argv);
+	if (rc != HB_MC_SUCCESS) { 
+		bsg_pr_err("failed to initialize grid.\n");
+		return rc;
+	}
+
+
+        /**********************************************************************/
+	/* Launch and execute all tile groups on device and wait for finish.  */ 
+        /**********************************************************************/
+	rc = hb_mc_device_tile_groups_execute(&device);
+	if (rc != HB_MC_SUCCESS) { 
+		bsg_pr_err("failed to execute tile groups.\n");
+		return rc;
+	}	
+
+
+        /**********************************************************************/
+        /* Freeze the tiles and memory manager cleanup.                       */
+        /**********************************************************************/
+	rc = hb_mc_device_finish(&device); 
+	if (rc != HB_MC_SUCCESS) { 
+		bsg_pr_err("failed to de-initialize device.\n");
+		return rc;
+	}
+
+	return HB_MC_SUCCESS;
+}
+
+#ifdef COSIM
+void test_main(uint32_t *exit_code) {	
+#ifdef VCS
+	svScope scope;
+	scope = svGetScopeFromName("tb");
+	svSetScope(scope);
+#endif
+	bsg_pr_test_info("test_tile_info Regression Test (COSIMULATION)\n");
+	int rc = kernel_tile_info();
+	*exit_code = rc;
+	bsg_pr_test_pass_fail(rc == HB_MC_SUCCESS);
+	return;
+}
+#else
+int main() {
+	bsg_pr_test_info("test_tile_info Regression Test (F1)\n");
+	int rc = kernel_tile_info();
+	bsg_pr_test_pass_fail(rc == HB_MC_SUCCESS);
+	return rc;
+}
+#endif
+
